backend/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
sqlalchemy==2.0.47
pyjwt==2.11.0
pwdlib==0.3.0